Clean up empty_line
Several different tokens are ignored by the parser -- INDENT (on a
line by itself, i.e. at EOF), SKIPPED, EOL, and COMMENT (when not part
of a transaction). There's no reason for SKIPPED to be separate from
the others (in the directive production). There's also no reason to
list possible sequences of them, because all are permitted (even if
they are impossible) and empty_line can be fired multiple times in
succession to accommodate them.