Fix #12
The previous code was showing the 64 oldest items, instead of the 64
newest items.

readline.get_history_item(0) is the oldest item in the readline
history. So if we want to show there are 200 items in the history
and we want to show the 64 newest we need to start iterating from
item (200-64) and show items 136-200.