Fix #311: Bypass MSVC quirk; missing strfmtime formats
For `#define STRINGIFY(s) #s` Visual Studio's preprocessor does
    `STRINGIFY()` -> <NOTHING>
and not the desired
    `STRINGIFY()` -> `""`

As a workaround, pre-insert an empty string if we detect we're in
MSVC-land, and let the compiled remove it.  If the stringified
argument is empty this prevents a syntax error.

The other problem: Windows's implementation of `strfmtime` apparently
lacks the `%s` format code used to get a timestamp.  Switch to
trimming the timezone part in Python, which should be cross-platform.