Fix running the booking_full algorithm with data from create_simple_posting().
The booking algorithm is primarily run with data produced by the parser, which
creates Postings that can be booked. But when starting with entries that are
produced in Python and data.create_simple_posting(), the resulting postings
may be missing fields (units and meta). Tolerate these fields being empty
when booking.

Fix #357