Permit (and ignore) lines starting with a colon.
Emacs's Org-mode, with which the Beancount syntax is carefully designed to be
compatible, has a feature called "drawers" that, among other things, allow for
certain subheading-level Emacs configuration to be specified. Drawers
are written with lines starting with a colon, for example the sequence

:PROPERTIES:
:visibility: all
:END:

immediately below a heading will cause Org-mode to leave that heading (but not
others) expanded when the file is first loaded.

These lines are easily identified as not relevant to Beancount, since the
syntax requires every line to start with either a date or a reserved word.
Other Org-mode constructs, such as headers (lines starting with * or #) and
comments (lines starting with ;) are treated as comments (i.e., ignored) by
the beancount lexer. This change makes lines starting with : also be ignored
by the lexer.