parser: Step 3 toward a reentrant parser
Provide a Parser Python object that encapsulates the scanner object.
Pass this object and the builder object to the lex and parse functions
generated by Flex and Bison. Move the file name and line number offset
from the scanner extra slot to the parser object. Keep the filename as
a python object, avoiding to have to decode and encode it all the time
and side stepping the issue of deciding which encoding to use for
doing so (I think that whas was done before was not exaclty correct).