parser: Simplify error handling in the lexer
Modify the calling convenction for methods of the Builder class that
construct token values to either return valid data or raise an
exception. It is responsability of the caller to catch the exception
an log the error.

This requires to tighten parsing somehow, returning an error in two
circumstances in which previously an error was reported but a value
produced anyway:

- string exceeding the maximum number of lines (64),
- numbers with misplaces comma thousands separators.