parser: Simplify error handling in the lexer
Modify the calling convention for methods of the Builder class that
construct token semantic values from the lexer. These methods must
return valid data or raise an exception. The responsibility to catch
exceptions and log errors is moved to the caller.

This requires to tighten parsing. An error is returned in two
circumstances in which previously an error was reported but a value
produced anyway:

- string exceeding the maximum number of lines (64),
- numbers with misplaces comma thousands separators.

This will allow to do not expose the scanner input location to the
builder class. This is a preparatory step to make the parser
reentrant.