parser: Modify builder methods calling convention
Most moethods used to build the parsing results need to access the
parsing location either to store it in the produced objects or for
error reporting. Modify the calling convention so that all Builder
methods called from the parser in the grammar are called with filename
and lineno as the first two arguments. This avoids the need for the
Builder to have access to the parser to fetch location information.

Rename the macro used in grammar.y from BUILD to CALL to reflect the
different calling convection used in lexer.l.