parser: Further decouple scanner and parser
Extend Bison default location tracking structure to include filename
and extend Flex location update facility to take into accoint file
name and line offset.  In this way the parser does not need to access
either the scanner or the Parser object.  Pass filename and line
offset to the scanner via the user extra data in the scanner object.
This allows to remove the Parser object from the scanner and parser
arguments.

Make sure that the same Parser object can be used to parse multiple
input streams resetting the scanner state in subsequent invocations.