beancount.el: Add beancount-electric-currency
This variable controls the behavior of RET. When non-nil, RET at the
end of a posting tries to add the currency to the amount, if
missing. All ways in which an amount can be specified are not
recognized yet, and the lookup for the account currenclyu can be
expensive in large ledger files. Thus do not enable this facility by
default.