parser: Simplify lexer handling of string literals
Simplify the lexer rules to do not require conditionally activated
rules to recognize string literals. Move handling of the C-like escape
sequences to easier to extend action code. This also allows to let
Flex handle the buffering and thus remove double buffering of the
string literals content (and some ugly code). As an extra benefit, the
lexer now reports the complete matched text for string literals. Tests
need to be adjusted accordingly.