parser: Represent lexer matched text as bytes and not strings
The input to the lexer is binary, thus this seems more correct and
avoid having to deal with text encoding issues (imprtant for chnages
to the literal strings handling coming in the next commit). The
matched text is used only in tests and for debugging, thus this change
should not have much of an impact, other than having to adjusts the
tests.