parser: Extend location tracking
Extend Bison default location tracking structure to include filename
and extend Flex location update facility to take into accoint file
name and line offset.  Pass filename and line offset to the scanner
via the user extra data in the scanner object.

In this way the grammar does not need to access to the scanner to
retrieve correct location information. This is a preparatory step to
make the parser reentrant.