parser: Store parsed file name as a Python object
The file name is nver used directly in the C code, thus doing this
avoids to convert the file name back to a Python string every time a
Builder method is called. This also sidesteps the problem of using the
right encoding to convert the file name to bytes and back. What has
been done so far is not completely correct:

https://docs.python.org/3/c-api/unicode.html#file-system-encoding