plugings.check_commodity: Do not depend on commodties set
This plugin is the only user of the set tracking all the commodities
encoutered in the parsing of the ledger stored in the options
dictionary. Move the detection of the commodities used in the ledger
to the plugin.

Scanning the ledger content for commodities is surprisingly fast: it
takes less than 20 ms on a 5 years ledger generated by bean-example.