parser: Use lookahead operator to simplyfy metadata keys matching
Using the Flex lookahead oeprator allows to simplify the action for
matching metadata keys. In turn this will allow to simplify the
token() macro removing the conditional return statement (and thus the
need for it to be a macro at all): better form and simpler code.

This also changes the matched text reported by the lexer and few tests
need to be adapted. I think the new form is more correct than the old.