Fix the CI
The most recent Ununtu release supported by Travis-CI is Focal which
ships with Bison 3.5. We now require Bison 3.6. This worked because
the Makefile decided not to rebuild the checked in generted files when
the commit requiring Bison 3.6 was introduced. When the Makefile
attempts to regenerate the grammar it fails.

Install a more recent Bison as part of the CI setup.