parser: Port the scanner generation from Flex to RE/flex
These are the minimal changes to the code required to plug a RE/flex
generated scanner into the existing parser. RE/flex can only generate
C++ code thus the main changes revolve around adapting to che
different calling conventions and to link the C++ code with the C
grammar.

The only change in the scanner rules is related to the definition of
account names. I was not able to tweak the old one so that RE/flex
would accept it, thus I went ahead and replaced it with the correct
Unicode aware one.