parser: Add tests for line number tracking
Also move the parse_doc() test from grammar tests to parser tests where
it belongs and reowrk it to use a different method to get the line
number is the source file. This avoid using basically the same method
to get the line number in the test and the implementation, thus
boosting confidence that the implementation is correct.