parser: Rework line number tracking
This avoids to have to store an offset line number in the lexer. At
the same time, revisit the handling of the report_firstline parameter
to make it do what it says. Previusly it was interpreted as an offset
to add to the scnner line number. Now it really specifies the line
number of the first input line, as the name suggests. Fix the
implementation of parse_doc() to follow this convention.