core.data: create data classes directly with NamedTuple
This should fix the issue of pylint not recognising all these classes as
callable, apparently the function creating and returning a NamedTuple
confused it too much.

This also fixes the typing of these central data structures, which was
quite incorrect. `new_directive` claimed to return an instance (of
NamedTuple) without any further typings but indeed returned a class. At
least with mypy, the typings further down the file all produced errors.

For these central data structures, I think the more explicit (and in my
eyes simpler) definition trumps the succinctness of not having to repeat
the shared first two fields meta and date.