Fixed #584 - a bug which was occurring because a pathological example using a very large precision was used in inferring the precision of a currency and a subsequent attempt to quantize a large number exceeded the default precision.
I fixed it by using a local context allowing for numbers as large as 1B without
problems. I think this may be slowing down the rendering, but I'm not sure, will
have to assess separately, and this will be done in v3 in the context of the
query engine rewrite.