cparser: Converted string symbol types to from std::string_view to std::string.
This was causing the occasional parsing bug (but rarely) due to buffer lifetime
rolling over during scanning. You have to copy. So we copy, and move all the way
through to the receiving protos to minimize cost. Measured the difference and it
isn't much (more variance, probably due to increased malloc()'ing).

I converted the date parsing to a fixed YYYY-MM-DD (there was a bug). Fixed
corresponding bug.

Made the unit tests for the scanner store the parsed and unescaped string value
contents instead of the matched strings (it's nicer).