plugins: Added an ignore regexp option to the check_commodity plugin.
This is useful if you have options and futures contracts. A lot of those symbols
are ephemeral. I added an option to narrowly ignore symbols in the context of
particular accounts.

You can provide a mapping of (account-regexp, currency-regexp) as options, to
specify which commodities to ignore from this plugin selectively. Use this
sparingly, as it is an out from the checks that this plugin provides. However,
in an active options trading account, a ton of products get inserted and the
number of commodity directives can be overwhelming and it's not productive to
declare each of the options contracts - names with an embedded strike and
expiration date, such as 'SPX_121622P3300' - individually.

Note that if a symbol has been ignored in at least one account, it will
therefore be further in all Price directives and Metadata values.