printer: Ignore all metadata fields whose key starts with "__"
Fields with key "__automatic__" are already ignored. Extend this to
all field with a key starting with "__". This allows to to use such
entries for attaching extra information to the entries, without having
to explicitly remove it before serializing the entries to file. The
immediate use is for markign duplicate entries identified during the
import process with a "__duplicate__" metadata field.

The "filename" and "lineno" metadata fields could be renamed to obey
to this scheme too and the special handling removed, but this requires
modifying the parser.