cache: Fix head() decoding
Fix file head byte data decoding when the byte string read from the
file terminates in the middle of an unicode character represented by
more than une byte. This is likely to occur for vaiable width
encodings such as utf8 with languages that uses characters beyond the
ascii set.

Fixes #645.