parser: Stop pretending we support encodings other than UTF-8
The parser does not really understand any encoding other than UTF-8.
Support for other encodings is limited to string literals, which makes
sense only for a very limited subset of encodings.

The limited support for other encodings is not exposed to the command
line tools and no one ever complained, so I assume that all Beancount
users use systems with UTF-8 default encoding or arrange for their
ledgers to be encoded in UTF-8 or in an encoding which is subset
equivalent to UTF-8 like ascii or latin1.