loader: Allow including the same file multiple times
Probably there are not many occasions where it is useful, but there
are no good reasons why it should not be allowed to include the same
file multiple times.

This just removes the tests that checks that including the same file
multiple times fails. The new include directive implementation will
allow it.