loader: Raise an exception on non existing input file
There are no good reasons why trying to load an non existent file
should result in a soft error. The command line tools anyway validate
the presence of the input file before we load them.

This jsut removes the tests that enforce this behavior. The actual
change will come with the refactoring of the load code.