Added an --auto/-a flag to bean-check.
This directive adds the "beancount.plugins.auto" plugin implicitly.

This is something I've been wanting for *years*. I try hard to minimize the
number of flags to Beancount programs but this one is essential: in many
context, one generates directives from programs and wants to evaluate this
output outside the context of the full file they are produced to be included in.
Typically, accounts and commodity declarations are missing. It is really useful
to be able to quickly evaluate a file like that without having to modify it
first.