Added a new "check_drained" plugin to verify that closed accounts are empty.
Hi!

I hate to break my own rules, but I've added a new plugin (though from an old
idea) to v2. The new plugin "check_drained" will insert zero balance checks for
all currencies seen in an account just before a Close directive. This should in
theory be part of the pedantic plugins but I won't change that in v2 (minimize
behavior changes). It will be in the pedantic list in v3.

Background: I finally had a bit of time to update/review my own
accounting (thanks to paternity leave) and had a bit of a "wtf!" moment when I
realized that this basic check was not included in Beancount. I'd committed the
sin of closing non-empty accounts in the past. I figured that since this plugin
really is core functionality and did not introduce any extra dependencies, I'd
include it in the frozen v2.