printer: do not sort metadata key upon printing
This makes the metadata key ordering fall back to the "natural" dictionary
ordering that, sarting with Python 3.7, is insertion order (as it was the case
for OrderedDicts with Python < 3.7).

As part of this change we hence also bump up Python required version to 3.7, so
that Beancount importers can impose a desired metadata ordering.

Closes #420.