Started updating regression comparison script (unfinished).
I couldn't finish this because we have to switch to the pyext API in order to be
able to convert to text with the `as_utf8` flag, and in order to do that I have
to use `text_format`, and that requires the pyext module, not the experimental
Bazel bindings one. I'll do that soon.