Fix: allow open and close directives on the same day
Previously, open and close directives on the same day caused an error.
However, Beancount considers opens to have occured on the beginning of a
day, and a close directive to have occurred at the end of a day.

This makes most sense, and was likely the original intention, as can be
evidenced by the fact that one can currently issue a transaction on the
same day as either a close, or an open, without errors. If so, issuing
an open and a close on the same day should also be valid. This fix
enables that.

See discussion here:
https://github.com/beancount/beancount/pull/751