parser: Remove line count limit from strings
The idea behind enforcing the limit was to detect instances where the
user forgot a closing quote. However, for how a beancount ledger is
usually structured, it is very hard to hit the line count limit before
the ledger contains an unrelated quote character that get interpreted
as the closing quote.

There are multiple reports about the limit being to low to for valid
input, for example in the case of long plugin configuration strings,
and the limit is not effective in detecting invalid ledgers. Remove
the limit.