Correct order of beancount.ops.pad in loader
The present language syntax specifies the following for the pad
directive:

> A padding directive automatically inserts a transaction that will make the subsequent balance assertion succeed, if it is needed

However, the present padding behaviour does not satisfy this. This issue has
been raised in the google groups at
https://groups.google.com/g/beancount/c/d4AaX1HqXwI.

This commit adds a test to the padding test already present in the
repository showing this behaviour. The added test uses a tempfile so I'm
not sure if cleanup is needed, however it is similar to the tempfile
used in loader_test.py.

This commit also adds the proposed change to loader.py which should
correct the behaviour. With the proposed tests, all the tests pass.

Should fix #742, but the forecast plugin is no longer in the repository.