build: Search win_flex and win_bison as alternatives for flex and bison
The WinFlexBison package https://github.com/lexxmark/winflexbison/ is
probably the easiest way to install flex and bison on Windows.
However, this packages installs flex and bison as win_flex and
win_bison respectively. Fall back to these if flex or bison are not
found under the canonical names. This makes compilation on Windows a
bit easier.