parser: Fix transaction directive round-trip test on Windows
beancount.parser.printer.print_entries() only escapes double quotes in
the transaction payee or narration. Escaping double quotes is
important because otherwise the resulting ledger would be invalid.
Escaping other special characters is responsibility of the user.
However, this means that attempting round-trip tests with special
characters is going to produce undesired results. Avoid doing so.

In particular, the specific implementation of the test was failing on
Windows because both \r and \n are interpreted as newline characters
when writing the ledger because of universal newlines. Thus the
transaction narration grew one line longer in the round-trip.

Fixes #875.