printer: Restore output format
The output format, including whitespace should be stable, as it is
used for example to test the output of importers in the beangulp
framework. This restores the output format to the one used before the
last commit. The change in trailing whitespace in the source location
comment is kept as that is a seldom used debug feature.

Adjust one test to allow the expected output to do not contain
trailing whitespace.

This partially reverts commit 82586a4f7888.